+++ email.php
@@ -62,11 +62,10 @@ $alternates = $DB->get_records_menu('block_clampmail_alternate',
     $altparams, '', 'id, address');
 
 $blockname = get_string('pluginname', 'block_clampmail');
-$header = get_string('email', 'block_clampmail');
+$header = get_string('composenew', 'block_clampmail');
 
 $PAGE->set_context($context);
 $PAGE->set_course($course);
-$PAGE->navbar->add($blockname);
 $PAGE->navbar->add($header);
 $PAGE->set_title($blockname . ': '. $header);
 $PAGE->set_heading($blockname . ': '.$header);
@@ -113,12 +112,6 @@ if ($groupmode == SEPARATEGROUPS && !has_capability('block/clampmail:cansendtoal
             unset($users[$userid]);
         }
     }
-}
-
-// Stop execution if there's no valid email target.
-$returnurl = new moodle_url('/course/view.php', array('id' => $course->id));
-if (empty($users)) {
-    notice(get_string('no_users', 'block_clampmail'), $returnurl);
 }
 
 if (!empty($type)) {
@@ -312,14 +305,24 @@ if (empty($warnings)) {
 
 echo $OUTPUT->header();
 echo $OUTPUT->heading($blockname);
+echo block_clampmail\navigation::print_navigation(
+        block_clampmail\navigation::get_links($course->id, $context),
+        get_string('composenew', 'block_clampmail')
+);
 
-foreach ($warnings as $type => $warning) {
-    $class = ($type == 'success') ? 'notifysuccess' : 'notifyproblem';
-    echo $OUTPUT->notification($warning, $class);
+// Don't show the form if there's no valid email target.
+$returnurl = new moodle_url('/course/view.php', array('id' => $course->id));
+if ($form->get_user_count() == 0) {
+    notice(get_string('no_users', 'block_clampmail'), $returnurl);
+} else {
+    foreach ($warnings as $type => $warning) {
+        $class = ($type == 'success') ? 'notifysuccess' : 'notifyproblem';
+        echo $OUTPUT->notification($warning, $class);
+    }
+
+    echo html_writer::start_tag('div', array('class' => 'no-overflow'));
+    $form->display();
+    echo html_writer::end_tag('div');
+
+    echo $OUTPUT->footer();
 }
-
-echo html_writer::start_tag('div', array('class' => 'no-overflow'));
-$form->display();
-echo html_writer::end_tag('div');
-
-echo $OUTPUT->footer();